Generates Mandelbrot Set image

+++ src/Mandelbrot.cpp
@@ -12,12 +12,12 @@ Mandelbrot::~Mandelbrot() {
     //Empty
 }
 
-int Mandelbrot::getIterations(double x, double y) {
+size_t Mandelbrot::getIterations(double x, double y) {
 
     complex<double> z = 0;
     complex<double> c(x, y);
 
-    int iterations = 0;
+    size_t iterations = 0;
 
     while(iterations < MAX_ITERATIONS){
         z = z*z + c;
